+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # `gitstat` Changelog
 
-## [0.1] (unreleased)
+## [0.1] UNRELEASED
+
+### Added
+
+* Support checking repos that are not tracked by gitstat
+* Add a fair amount of type hinting
+
+### Changed
 
 * Use 'Click' to manage command-line options